Updated file structure and reformatted LinkOwnerTypeController to eslint standards
