+++ server/src/Server.ts
@@ -1,8 +1,9 @@
 import App from "./App";
-import UsersController from "./api/controllers/UsersController";
-import GroupsController from "./api/controllers/GroupsController";
-import ResourceTypeController from "./api/controllers/ResourceTypeController";
-import ResourceController from "./api/controllers/ResourceController";
+import UsersController from "./controllers/UsersController";
+import GroupsController from "./controllers/GroupsController";
+import ResourceTypeController from "./controllers/ResourceTypeController";
+import ResourceController from "./controllers/ResourceController";
+import LinkOwnerTypeController from "./controllers/LinkOwnerTypeController";
 // Creates a new Express App that takes a list of Controllers and a port
 const app = new App(
   [
@@ -10,6 +11,7 @@ const app = new App(
     new GroupsController(),
     new ResourceTypeController(),
     new ResourceController(),
+    new LinkOwnerTypeController(),
   ],
   8080
 );
