+++ server/src/controllers/ResourceController.ts
