+++ server/src/models/CategoryTag.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model } from "sequelize";
-import sequelize from "../databaseConnection";
+import sequelize from "../middleware/databaseConnection";
 
 import Tag from "./Tag";
 import Category from "./Category";
