+++ server/src/models/GroupCategory.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model } from "sequelize";
-import sequelize from "../databaseConnection";
+import sequelize from "../middleware/databaseConnection";
 import Group from "./Group";
 import Category from "./Category";
