+++ server/src/models/Resource.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model } from "sequelize";
-import sequelize from "../databaseConnection";
+import sequelize from "../middleware/databaseConnection";
 
 import User from "./User";
 import ResourceType from "./ResourceType";
