+++ server/src/models/ResourceVersion.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model } from "sequelize";
-import sequelize from "../databaseConnection";
+import sequelize from "../middleware/databaseConnection";
 
 import Resource from "./Resource";
 import Link from "./Link";
