+++ server/src/models/User.ts
@@ -1,5 +1,5 @@
 import { DataTypes, Model } from "sequelize";
-import sequelize from "../databaseConnection";
+import sequelize from "../middleware/databaseConnection";
 import bcrypt from "bcrypt";
 
 class User extends Model {
